fix(ui): the settle sidecar stands down where the deputy brokers the script (#695)
`todo add-repeating --dangerously-drive-gui` has been failing in ~2 s on every
helpers-routed Mac since 0.20.7, with nothing driven:

    script rejected: contains "do shell script" — the deputy only brokers
    GUI/AppleEvent scripts, never shell execution

Both of the AX settle sidecar's hops shell out — the spawn hop and the
per-settle `printf | nc -U` client — and the deputy's broker refuses that
phrase on purpose (`scriptGuard`), so that "drive the Things GUI" can never
become "run arbitrary shell under the helper's grants". The lint is correct and
stays; the generator yields.

Deputy routing is now one more `observerAvailable()` = false reason
("deputy-routed: brokered scripts cannot spawn the sidecar", traced like every
other unavailability reason), which regenerates every script byte-identically
to the polling version VOPAT2 certified as its fallback. The observer stays
live on every direct-execution host and nothing else in the drive changes. The
gate asks whether the deputy is EXPECTED rather than merely active: an
expected-but-dead deputy already makes the osascript seam refuse (#620), so no
configuration runs an acting script locally while expecting one.

Why nothing caught it: goldens have no helpers, so every lab arm ever run
executed its scripts DIRECT. "Which identity executes this script" was a
dimension with one value, and the campaign predicted the missing one would work
by construction from a sound identity law — true of a process the deputy
spawns, except the deputy never runs the script that spawns it.

Three guards, so the class cannot ship again:

- the banned list is ONE TS constant (DEPUTY_BANNED_SCRIPT_PHRASES) pinned to
  the Swift array by a drift test that parses deputy/src/server.swift
- test/unit/ui-script-broker-safety.test.ts renders every acting script the ui
  vector can emit in the shape a routed host generates and asserts none carries
  a banned phrase, with a negative control proving the observed shape IS what
  the broker refuses
- the ROUTED arm: test/deputy/broker-integration.test.ts now feeds the same
  catalog to the real Swift broker binary (macOS CI; ran green locally, 14/14)

The script catalog is shared with the osacompile suite
(test/unit/helpers/ui-script-catalog.ts) so a script cannot be known to one
guard and invisible to the other.

Still a gap, recorded and queued: a COMPLETE routed GUI drive. Installing the
helpers in a clone needs a Swift toolchain, a Developer ID signature and
interactive TCC consent inside an airgapped guest — its own campaign. Until
then it is covered by the release gate now on record: a field-shaped
end-to-end on a real routed host with a real display is REQUIRED before any
tag, in addition to lab:regress.

Also queued as a decision, NOT built: the observer belongs in the trusted
long-lived process — a deputy-hosted AXObserver, or a deputy verb that spawns
the sidecar. Either is a deputy version bump plus a helpers reinstall;
maintainer's call.

Refs #695
