+++ docs/design/decisions.md
@@ -315,3 +315,5 @@ One line per ruling: **date · ruling · evidence**. This register exists to pre
 - 2026-09-02 · **An after-completion deadline offset at or above the series' own period is REFUSED before dispatch, because the app enforces its cap by silent substitution (DEFAULTS2, oddities §32).** Measured across 6 seed offsets × 6 unit/interval pairs: under `after completion, every N <unit>` the dialog caps `and start N days earlier` at `P − 1` days (a month counted as 30, a year as 365), since a start further back would fall on or before the previous occurrence's due date. That constraint is coherent; its ENFORCEMENT is not — a typed 30 became 6 under `every 1 week` and 0 under `every 3 days`, with no refusal and no sign, and the committed rule carried the substitute. The drive would have caught it twice over (the pre-commit audit re-reads the field; the post-drive oracle asserts the landed `startOffsetDays` against the requested one) but both of those fail a request that never had a chance, so the over-caution fail direction says refuse it and name the cap. This also closes a latent defect: `--after-completion --frequency weekly --interval 1 --start-days-earlier 7` used to be accepted and silently landed `ts=-6`. Evidence: [lab/defaults2-minimal-recipe.md](../lab/defaults2-minimal-recipe.md) §2.
 
 - 2026-09-03 · **The pre-release `lab:regress` gate is REAL, and it is BROADENED: a release is certified by a DEPUTY-ROUTED, normal-CLI, real-display run, not by lab conditions (maintainer ruling).** The written gate ([reference/suite-audit.md](../reference/suite-audit.md) § adopted rule) was ignored for five releases and honored on every release since 2026-08-28; the open "rule vs practice — which bends?" question is closed in the rule's favor. A RED `lab:regress` blocks the tag, AND a GREEN one is no longer sufficient: every operation whose driver / vector / recipe / deputy-facing code changed in a batch must be run end-to-end at least once **through the deputy** (helpers installed and enabled) using the **normal CLI syntax** against synthetic content, and every GUI-driven (ui-vector) one at least once on a **real display**. Lab probes stay direct-execution by design — that is what they are for — but release certification is the field-shaped path. Evidence: two lab-green/field-red releases in five weeks, same root cause. **v0.19.2** — the window/focus census opened with an unaddressed system-wide process enumeration and stalled ~15 s per inspection on the maintainer's host, failing the drive with nothing typed; timed against a standing sheet in a clone it returned in ~60–250 ms and the stall did not reproduce headless at all ([#629](https://github.com/mikegreiling/things-api/issues/629), [lab/fgrd2-census-hardening.md](../lab/fgrd2-census-hardening.md)). **v0.20.7** — #687's observer sidecar reached its socket through `do shell script`, which the deputy's `scriptGuard` (`deputy/src/server.swift`) refuses by design, so it was green on every lab probe and broken on every helpers-enabled host (hotfix branch `mg/observer-deputy-routing`). The rule lives in [AGENTS.md](../../AGENTS.md) § Conventions → *Release gate*; the drill it gates is [reference/release-checklist.md](../reference/release-checklist.md). The automatable half — helpers installed inside a lab guest, so a routed arm can run headless — is queued in [up-next.md](../up-next.md); the real-display half stays human regardless.
+
+- 2026-09-03 · **A generated script must be BROKERABLE, and the deputy's ban list lives in exactly one place per language (#695).** The AX settle sidecar reaches its socket through `do shell script`, and the deputy's broker refuses that phrase on purpose (`scriptGuard`) — so 0.20.7 shipped a `todo add-repeating --dangerously-drive-gui` that failed in two seconds on every helpers-routed Mac, the maintainer's own included, while every lab arm stayed green because goldens run scripts DIRECT. Three rulings. **(a)** The lint stays; the GENERATOR yields. Deputy routing is one more `observerAvailable()` = false reason (`deputy-routed: brokered scripts cannot spawn the sidecar`), which regenerates every script byte-identically to the certified polling version — availability failure is a fallback, never a refusal, and this is one more availability failure. The gate asks whether the deputy is EXPECTED, not merely active, because an expected-but-dead deputy already makes the osascript seam refuse (#620), so no configuration runs an acting script locally while expecting one. **(b)** The ban list is ONE TS constant (`DEPUTY_BANNED_SCRIPT_PHRASES`) pinned to the Swift array by a drift test that parses `deputy/src/server.swift` — two hand-kept lists in two languages drift silently until a field command fails. **(c)** WHICH IDENTITY EXECUTES A SCRIPT is a certification dimension: the ui script catalog is checked against the mirror on every `npm run check` and against the real Swift broker binary in the live-deputy suite, and no lab verdict may say "certified" without naming the execution identity it was certified under. Evidence: [lab/vopat2-screen-reader-build.md](../lab/vopat2-screen-reader-build.md) §10, [lab/harness.md](../lab/harness.md) §Routing-arm law.
