+++ docs/lab/harness.md
@@ -156,7 +156,13 @@ AX-driven surfaces are **undocumented private APIs**: no contract, no deprecatio
 - **Never require a notification you have not seen fire through the registration you will use.** `AXLayoutChanged` never fires at all (VOPAT1-12); and the `AXUIElementDestroyed` burst that VOPAT1 saw for a cadence-group rebuild does NOT arrive at an APPLICATION-element registration (VOPAT2-4), so a settle that required it hung for its whole budget. Registration scope is part of the measurement.
 - **Removing a wait is a change to a DEPENDENCY, not to a duration** (VOPAT2-7). Twice in one build, deleting a settle that "obviously" had nothing to wait for simply moved the cost: a value-unchanged pop-up selection whose settle was skipped outright left the NEXT hop's click to be swallowed by an app still closing its menu (1,515 ms, then a retry that worked in 4.1 ms). The only safe replacement for a wait is a POSITIVE observable for whatever it was actually holding — never nothing, and never a shorter clock. This is RDLAT2's settle law from the other side, and it will recur.
 
-**Sidecar law: an observing process is bounded on three independent axes.** The observer cannot live inside an osascript hop (`AXObserverCreate` takes a C function pointer the JXA bridge will not marshal), so it is a `python3` ctypes process — and it is spawned from INSIDE the hop, which is what makes it inherit the Accessibility identity that already holds the grant (the deputy on a helper-routed Mac, the terminal otherwise: responsibility is spawn-assigned, APDP1) and is why no new deputy verb and no new consent were needed. Because a process that outlives what it observed is a bug, it is stopped in the drive's `finally`, AND it exits on an absolute TTL, AND it exits after a no-request idle window. Certify all three, plus a "no stray pids" check after every drive cell.
+**Sidecar law: an observing process is bounded on three independent axes.** The observer cannot live inside an osascript hop (`AXObserverCreate` takes a C function pointer the JXA bridge will not marshal), so it is a `python3` ctypes process — and it is spawned from INSIDE the hop, which is what makes it inherit the Accessibility identity that already holds the grant (the terminal, on a direct-execution host: responsibility is spawn-assigned, APDP1) and is why no new consent was needed. Because a process that outlives what it observed is a bug, it is stopped in the drive's `finally`, AND it exits on an absolute TTL, AND it exits after a no-request idle window. Certify all three, plus a "no stray pids" check after every drive cell. **It cannot exist on a deputy-routed host at all** — see the routing-arm law below, which is the correction this law originally got wrong.
+
+**Routing-arm law: WHICH IDENTITY EXECUTES A SCRIPT is a certification dimension, and every arm the lab has is one value of it** (#695, field regression in 0.20.7). Goldens have no helpers installed, so every drive the lab has ever certified ran its scripts DIRECT, under the guest's own terminal. A helpers-routed host is a different execution path with a different gate on it: the deputy LINTS every script it brokers and refuses any containing `do shell script` or `do script` (`scriptGuard`, deputy/src/server.swift — mirrored in TS as `DEPUTY_BANNED_SCRIPT_PHRASES`), because a broker that will shell out is no longer "drive the Things GUI" but "run arbitrary shell under the helper's grants". The AX settle sidecar's two hops both shell out, so it shipped fully certified and failed in two seconds on every routed Mac — including the maintainer's own. Three rules:
+
+- **A green lab arm is a verdict about DIRECT execution only.** Never write "certified" without naming the identity it was certified under. A vector whose scripts change shape must be re-asked under both.
+- **The routed arm the lab cannot host, CI can.** Installing and granting the helpers inside a clone is not feasible quickly (the bundle needs a Swift toolchain and a Developer ID signature, and Accessibility + Automation grants for it need interactive consent, inside an airgapped guest whose whole doctrine is that nothing raises a dialog). The certifiable half is acceptance at the broker's DOOR, and it needs no GUI: `test/deputy/broker-integration.test.ts` builds the real Swift broker and now feeds it every acting script the ui vector generates, with a stub osascript behind it. A real routed GUI drive stays a gap, tracked as such.
+- **The field-shaped end-to-end is REQUIRED before a tag** — the release gate, canonized in [AGENTS.md](../../AGENTS.md) § Conventions and drilled in [reference/release-checklist.md](../reference/release-checklist.md). `npm run lab:regress` green is necessary and not sufficient: a candidate is also run once on a real routed host, with a real display for GUI drives, using the normal CLI syntax against synthetic content. A host class with no lab arm is a host class only the field can answer for, and this is the second release in five weeks to prove it (#629 was the first).
 
 **Verify-by-read law: an actuation whose value the surface already holds becomes a READ — but only a read that was actually taken** (DEFAULTS2, 2026-09-02). Where the driver MINTS the state a dialog seeds itself from, it can compute what the dialog will pre-fill before opening it, and every control whose pre-fill is already correct is a write that can become one read. Four rules, and the third is what keeps it honest:
