+++ src/write/vectors/ui-observer.ts
@@ -62,13 +62,35 @@
  *    an explicit `stop` in the drive's `finally`. A sidecar cannot outlive its
  *    drive even if the drive is killed.
  *
+ * ── ONE HOST CLASS CANNOT HAVE IT: the deputy-routed Mac (#695) ────────────
+ *
+ * Both shell hops above are the problem. The deputy's broker LINTS every script
+ * it is handed and refuses any containing `do shell script` (or `do script`) —
+ * `scriptGuard`, deputy/src/server.swift, mirrored in TS as
+ * DEPUTY_BANNED_SCRIPT_PHRASES — because a broker that will shell out is no
+ * longer "drive the Things GUI" but "run arbitrary shell under the helper's
+ * grants". That lint is correct and stays. What it means is that a host which
+ * routes its automation through the deputy cannot have this sidecar AT ALL: not
+ * the spawn hop, not the in-script client. 0.20.7 shipped the sidecar without
+ * that gate and `todo add-repeating --dangerously-drive-gui` failed in two
+ * seconds on every helpers-routed Mac — the lab never saw it because goldens run
+ * scripts DIRECT, with no deputy installed, so the certified arm was the only arm.
+ *
+ * So deputy routing is simply one more reason {@link observerAvailable} says no,
+ * and the fallback below does the rest. The observer stays live on every
+ * direct-execution host. The proper long-term home for an AXObserver is the
+ * trusted long-lived process itself — a deputy-hosted observer, or a deputy verb
+ * that spawns the sidecar under the helper's identity — which is a deputy
+ * version bump plus a helpers reinstall, and is the maintainer's call
+ * (docs/lab/vopat2-ax-settle-observer.md §routed arm).
+ *
  * ── the fallback stays certified ───────────────────────────────────────────
  *
  * `python3` ships with the Command Line Tools, so availability is gated exactly
- * as the probe gated it (`xcode-select -p`). When the gate says no — or the
- * sidecar cannot be spawned, or its handshake fails — every script is generated
- * BYTE-IDENTICALLY to the polling version that shipped before this campaign and
- * the drive proceeds on it. Availability failure is a fallback, never a
+ * as the probe gated it (`xcode-select -p`). When the gate says no — or the host
+ * is deputy-routed, or the sidecar cannot be spawned, or its handshake fails —
+ * every script is generated BYTE-IDENTICALLY to the polling version that shipped
+ * before this campaign and the drive proceeds on it. Availability failure is a fallback, never a
  * refusal. An armed settle that TIMES OUT is the opposite: it fails closed with
  * its own named reason, because a settle that gave up is a surface the drive
  * cannot vouch for (determinism doctrine, fail direction over-caution).
@@ -79,6 +101,7 @@ import { chmod, mkdir, readFile, writeFile } from "node:fs/promises";
 import { createHash, randomBytes } from "node:crypto";
 import { join } from "node:path";
 
+import { deputyExpected } from "../../deputy/routing.ts";
 import { stateDir } from "../../paths.ts";
 import { trace } from "../../trace/tracer.ts";
 import { escapeAppleScript } from "./applescript.ts";
@@ -720,11 +743,26 @@ function which(cmd: string, args: string[]): Promise<boolean> {
  * Command Line Tools are what put `python3` on the PATH, and `xcode-select -p`
  * is the prompt-free question that says whether they are installed. Memoized —
  * the answer cannot change inside one process.
+ *
+ * ONE GATE COMES FIRST, AND IT IS NOT ABOUT THIS MACHINE'S TOOLS (#695): a host
+ * that expects the deputy to carry its automation cannot have the sidecar,
+ * because both of the sidecar's hops shell out and the broker's lint refuses
+ * every script that does. EXPECTED, not merely active, is the right question:
+ * when the deputy is expected but not carrying traffic the osascript seam
+ * REFUSES rather than running direct (#620, src/deputy/osa.ts), so there is no
+ * configuration where an expected deputy still executes an acting script
+ * locally. The check is prompt-free and socket-free — config plus a file
+ * existence test — so asking it costs nothing and can never raise a dialog.
+ * It is deliberately OUTSIDE the memo: routing is a per-process decision of its
+ * own, and the test seams reset the two independently.
  */
 export async function observerAvailable(
   env: NodeJS.ProcessEnv = process.env,
 ): Promise<{ ok: boolean; why: string }> {
   if (observerDisabled(env)) return { ok: false, why: `switched off by ${OBSERVER_ENV}` };
+  if (deputyExpected(env)) {
+    return { ok: false, why: "deputy-routed: brokered scripts cannot spawn the sidecar" };
+  }
   if (availability !== null) return availability;
   if (!(await which("/usr/bin/xcode-select", ["-p"]))) {
     availability = { ok: false, why: "the Command Line Tools are not installed (xcode-select -p)" };
