+++ test/unit/ui-script-syntax.test.ts
@@ -32,259 +32,9 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { describe, expect, it } from "vitest";
 
-import {
-  makeRepeatingRecipe,
-  projectMakeRepeatingRecipe,
-  projectPauseRepeatRecipe,
-  projectRescheduleRepeatRecipe,
-  pauseRepeatRecipe,
-  rescheduleRepeatRecipe,
-  resumeRepeatRecipe,
-  type RepeatRuleExtras,
-} from "../../src/write/vectors/ui-recipes.ts";
-import type { RepeatDialogShape, UiStep } from "../../src/write/vectors/types.ts";
-import {
-  axAbortScript,
-  axAuditDialogScript,
-  axCancelDialogScript,
-  axCancelFrameScript,
-  axSetValueScript,
-  axTypeTextScript,
-  axVerifyPrefillDateAreasScript,
-  axVerifyPrefillScript,
-  commandForStep,
-} from "../../src/write/vectors/ui.ts";
-import {
-  inertSettleInjector,
-  type SettleInjector,
-  settleInjectorFor,
-} from "../../src/write/vectors/ui-observer.ts";
-import { axFocusGuardPrelude, axUiStateScript } from "../../src/write/vectors/ui-state.ts";
+import { everyUiScript } from "./helpers/ui-script-catalog.ts";
 
 const DARWIN = process.platform === "darwin";
-
-/** Resolve a recipe's steps for one dialog shape, exactly as `drive()` does. */
-function forShape(steps: UiStep[], shape: RepeatDialogShape): UiStep[] {
-  return steps
-    .filter((s) => s.onlyShape === undefined || s.onlyShape === shape)
-    .map((s) => (s.shaped === undefined ? s : Object.assign({}, s, s.shaped[shape])));
-}
-
-/**
- * The full deadlined vocabulary — every optional control the dialog has, so no
- * conditional branch of the recipe goes ungenerated.
- */
-const FULL: RepeatRuleExtras = {
-  weekdays: ["monday", "thursday"],
-  monthly: { day: 20 },
-  yearly: { month: 8, day: 20 },
-  ends: { kind: "after", count: 5 },
-  reminder: "09:00",
-  deadline: true,
-  startDaysEarlier: 14,
-  next: "2026-08-20",
-};
-
-/**
- * A candidate-addressed step compiles to its own in-script resolution prelude
- * plus the addressed body (DRVLAT1, issue #633), so nothing has to be bound here:
- * `commandForStep` alone yields exactly the script that is dispatched. This suite
- * therefore covers the FOLDED shape — prelude, deadline loop and all — which is
- * the only shape that ever reaches osascript.
- */
-
-function everyScript(): { label: string; script: string; lang: string }[] {
-  const recipes = [
-    makeRepeatingRecipe("T-1", "yearly", 2, FULL),
-    // THE SEEDED SHAPE (DEFAULTS2): the same vocabulary with a seed row whose
-    // dates PROVE the pre-fill, so the verify-by-read branch and the tagged
-    // setters are generated as well as the untagged ones.
-    makeRepeatingRecipe("T-1", "weekly", 1, {
-      weekdays: ["thursday"],
-      reminder: "09:30",
-      next: "2026-07-09",
-      seed: {
-        scheduled: "2026-07-09",
-        today: "2026-07-05",
-        deadline: null,
-        reminder: "09:30",
-      },
-    }),
-    makeRepeatingRecipe("T-1", "monthly", 1, { monthly: { weekday: "monday", ordinal: 2 } }),
-    makeRepeatingRecipe("T-1", "weekly", 1, { weekdays: ["sunday"], afterCompletion: false }),
-    makeRepeatingRecipe("T-1", "daily", 1, { afterCompletion: true }),
-    makeRepeatingRecipe("T-1", "daily", 3),
-    rescheduleRepeatRecipe("T-1", "yearly", 1, FULL),
-    projectMakeRepeatingRecipe("AREA-1", "P-1", "A project", "monthly", 1, FULL),
-    projectRescheduleRepeatRecipe("P-1", "weekly", 1, FULL),
-    pauseRepeatRecipe("T-1"),
-    resumeRepeatRecipe("T-1"),
-    projectPauseRepeatRecipe("P-1"),
-  ];
-  const out: { label: string; script: string; lang: string }[] = [];
-  const seen = new Set<string>();
-  // BOTH SETTLE SHAPES (VOPAT2). Every step compiles twice: once with no settle
-  // sidecar — the polling script that shipped before that campaign — and once
-  // with one live, where the same step carries the observer handlers and its
-  // `nc -U` waits. A settle snippet that does not parse would fail mid-dialog,
-  // which is precisely the class this suite exists for.
-  const injectors: { tag: string; obs: SettleInjector }[] = [
-    { tag: "polling", obs: inertSettleInjector() },
-    {
-      tag: "observed",
-      obs: settleInjectorFor({
-        socketPath: "/tmp/things-api-observer/s-0123abcd.sock",
-        token: "0123456789abcdef0123456789abcdef",
-        logPath: "/tmp/things-api-observer/observer.log",
-        registered: "16/16",
-        pid: 4242,
-      }),
-    },
-  ];
-  for (const recipe of recipes) {
-    for (const shape of ["next-popup", "legacy"] as const) {
-      for (const { tag, obs } of injectors) {
-        for (const step of forShape(recipe.steps, shape)) {
-          const cmd = commandForStep(step, recipe.targetUuid, obs);
-          if (typeof cmd.script !== "string") continue;
-          if (seen.has(cmd.script)) continue;
-          seen.add(cmd.script);
-          out.push({
-            label: `${recipe.op} · ${shape} · ${tag} · ${cmd.label}`,
-            script: cmd.script,
-            lang: cmd.lang ?? "applescript",
-          });
-        }
-      }
-    }
-  }
-  // The scripts the DRIVER compiles rather than the recipe: the read-only
-  // census, the cleanup ladder's two dismissals, and the pre-commit audit —
-  // whose occurrence comparison carries the relative-date resolver (#625).
-  for (const extra of [
-    { label: "driver \u00b7 ui-state census", script: axUiStateScript() },
-    // The FOLDED focus guard: the census prelude compiled in front of the very
-    // keystroke script it guards (DRVLAT1). Both dialog invariants, since the
-    // expected-sheet clause is generated only when one is latched.
-    {
-      label: "driver \u00b7 folded focus guard (no sheet latched) + set-value",
-      script: `${axFocusGuardPrelude(null)}\n${axSetValueScript("text field 1 of sheet 1 of window 1", "3")}`,
-    },
-    {
-      label: "driver \u00b7 folded focus guard (repeat sheet latched) + type-text",
-      script: `${axFocusGuardPrelude("repeat")}\n${axTypeTextScript("Some Project")}`,
-    },
-    { label: "driver \u00b7 dismiss (Cancel)", script: axCancelDialogScript() },
-    { label: "driver \u00b7 Cancel button frame", script: axCancelFrameScript() },
-    { label: "driver \u00b7 abort (Escape)", script: axAbortScript() },
-    {
-      label: "driver \u00b7 pre-commit audit",
-      script: axAuditDialogScript({
-        shell: "sheet 1 of window 1",
-        group: "group 1 of sheet 1 of window 1",
-        controls: [
-          {
-            kind: "popup",
-            label: "frequency",
-            path: "pop up button 1 of sheet 1 of window 1",
-            expected: ["daily"],
-          },
-          {
-            kind: "group-number",
-            label: "interval",
-            numberTarget: "interval",
-            expected: ["3"],
-          },
-          {
-            kind: "occurrence-popup",
-            label: "first occurrence",
-            path: "pop up button 2 of group 1 of sheet 1 of window 1",
-            expected: ["2026-08-20"],
-          },
-          {
-            kind: "checkbox",
-            label: "add reminders",
-            path: "checkbox 1 of sheet 1 of window 1",
-            expected: ["1"],
-          },
-          { kind: "weekdays", label: "weekdays", weekdayBase: 3, expected: ["Monday"] },
-          { kind: "row-field", label: "start earlier", rowLabel: "days earlier", expected: ["14"] },
-        ],
-      }),
-    },
-    // THE VERIFY-BY-READ LEGS (DEFAULTS2). Both are driver-compiled for the
-    // audit's reason — the plan's controls are addressed through the live shell
-    // and the measured shape — so neither reaches osascript via a recipe step,
-    // and both carry the relative-date resolver the #625 class needs.
-    {
-      label: "driver · verify-prefill (System Events leg)",
-      script: axVerifyPrefillScript({
-        shell: "sheet 1 of window 1",
-        group: "group 1 of sheet 1 of window 1",
-        controls: [
-          {
-            kind: "group-number",
-            label: "interval",
-            numberTarget: "interval",
-            expected: ["1"],
-            prefillKey: "interval",
-          },
-          {
-            kind: "popup",
-            label: "after-completion unit",
-            path: "pop up button 1 of group 1 of sheet 1 of window 1",
-            expected: ["week", "weeks"],
-            prefillKey: "ac-unit",
-          },
-          {
-            kind: "occurrence-popup",
-            label: "first occurrence",
-            path: "pop up button 2 of group 1 of sheet 1 of window 1",
-            expected: ["2026-08-20"],
-            prefillKey: "next",
-          },
-          {
-            kind: "checkbox",
-            label: "add reminders",
-            path: 'checkbox "Add reminders" of sheet 1 of window 1',
-            expected: ["1"],
-            prefillKey: "add-reminders",
-          },
-          {
-            kind: "weekdays",
-            label: "weekdays",
-            weekdayBase: 3,
-            expected: ["Monday"],
-            prefillKey: "weekdays",
-          },
-          {
-            kind: "row-field",
-            label: "start earlier",
-            rowLabel: "days earlier",
-            expected: ["14"],
-            prefillKey: "start-earlier",
-          },
-        ],
-      }),
-    },
-  ]) {
-    if (seen.has(extra.script)) continue;
-    seen.add(extra.script);
-    out.push({ ...extra, lang: "applescript" });
-  }
-  const jxa = {
-    label: "driver · verify-prefill (date-area leg)",
-    script: axVerifyPrefillDateAreasScript([
-      { label: "reminder", target: "reminder", spec: "time:09:30", prefillKey: "reminder-time" },
-    ]),
-    lang: "javascript",
-  };
-  if (!seen.has(jxa.script)) {
-    seen.add(jxa.script);
-    out.push(jxa);
-  }
-  return out;
-}
 
 /**
  * Does this script address the Things application ITSELF, rather than driving its
@@ -343,7 +93,7 @@ function compileError(script: string, lang: string, dir: string, n: number): str
 describe.skipIf(!DARWIN)("every generated AppleScript compiles (NEXTPOP1)", () => {
   it("osacompile accepts every script the repeat recipes emit", () => {
     const dir = mkdtempSync(join(tmpdir(), "ui-script-syntax-"));
-    const all = everyScript();
+    const all = everyUiScript();
     // A recipe set that generated nothing would pass vacuously.
     expect(all.length).toBeGreaterThan(30);
 
@@ -369,7 +119,7 @@ describe.skipIf(!DARWIN)("every generated AppleScript compiles (NEXTPOP1)", () =
   // The filter above must never become the reason the suite is green. On a host
   // WITH Things every script is checked, app-addressed ones included.
   it("checks the app-addressed scripts too wherever Things is installed", () => {
-    const appAddressed = everyScript().filter((s) => needsThingsDictionary(s.script));
+    const appAddressed = everyUiScript().filter((s) => needsThingsDictionary(s.script));
     expect(appAddressed.length).toBeGreaterThan(0);
     if (!thingsInstalled()) return;
     const dir = mkdtempSync(join(tmpdir(), "ui-script-syntax-app-"));
